// projects/mfe-contabilidad/src/app/core/dto/list-ccomproba-v.dto.ts
export interface ListCcomprobaVDto {
  ccoCodigo:      number;
  dspComproba:    string;
  fecha:          string;
  empresa:        number;
  periodo:     number;
  sigla:          number;
  almacen:        number;
  serie:       number;
  numero:      number;
  doctran:     string;
  tipodoc:     number;
  concepto:    string;
  modulo:      number;
  nocontable:  boolean;
  estado:      number;
  descuadre:   boolean;
  adestino:    number;
  pventa:      number;
  centro:      number;
  tipoCambio:  number;
  tclipro:     boolean;
  codclipro:   number;
  agente:      null;
  cuenta:      null;
  transacc:    number;
  codclipro1:  null;
  cieComproba: null;
  refComproba: null;
  anulado:     number;
  anuComproba: null;
  autTipo:     null;
  nivelAprob:  null;
  tpdId:          string;
  tpdNombre:      string;
  modCodigo:      number;
  modId:          string;
  modNombre:      string;
  ctiId:          string;
  ctiNombre:      string;
  ctiTipo:        boolean;
  ctiAutoriza:    boolean;
  almNombre:      string;
  almId:          string;
  fechafin:       null;
  mes:            number;
  cliId:          string;
  ccoNombre:      string;
  valComproba:    null;
}